backend/requirements.txt
alembic
annotated-types
anyio
asyncpg
bcrypt
cachetools
certifi
charset-normalizer
click
ecdsa
fastapi
google-auth
greenlet
h11
httpx
idna
jose
Mako
MarkupSafe
passlib
psycopg2-binary
pyasn1_modules
pydantic
pydantic_core
python-dotenv
python-jose
requests
rsa
six
sniffio
SQLAlchemy
starlette
supabase
typing-inspection
typing_extensions
urllib3
uvicorn

